requirements.txt
phidata
streamlit
zep-python
google-generativeai
python-dotenv
pydantic
